backend/requirements.txt
# Django and REST framework
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
django-cors-headers>=3.7.0,<4.0.0
django-environ>=0.4.5,<1.0.0
gunicorn>=20.1.0,<21.0.0

# Document processing
pymupdf>=1.18.0  # For PDF processing
python-docx>=0.8.11  # For DOCX processing

# NLP and vector embeddings
nltk>=3.6.0,<4.0.0
scikit-learn>=0.24.0,<2.0.0
spacy>=3.0.0,<4.0.0
sentence-transformers>=2.0.0  # For embeddings

# Database
psycopg2-binary>=2.9.1,<3.0.0  # PostgreSQL adapter
SQLAlchemy>=1.4.0,<2.0.0  # For advanced database operations

# Utilities
python-dateutil>=2.8.2,<3.0.0
pytz>=2021.1
tqdm>=4.61.0,<5.0.0

# Development tools
pytest>=6.2.4,<7.0.0
pytest-django>=4.4.0,<5.0.0
black>=21.5b2
flake8>=3.9.2,<4.0.0
